// src/payload-types.ts
export interface User {
  id: number;
  name?: string | null;
  /**
   * admin manages settings and users; editor manages content; viewer can read admin data.
   */
  roles?: ('admin' | 'editor' | 'viewer')[] | null;
  updatedAt: string;
  createdAt: string;
  email?: string | null;
  username?: string | null;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}